compiler warning cleanup
git-svn-id: http://svn.freeswitch.org/svn/freeswitch/trunk@2487 d0543943-73ff-0310-b7d9-9358b9ac24b2

+++ src/mod/codecs/mod_g726/mod_g726.c
@@ -193,7 +193,7 @@ static switch_status_t switch_g726_decode(switch_codec_t *codec,
 			switch_bitpack_out(&handle->unpack, in[z++]);
 		}
 		for(y = 0; y < handle->bytes; y++) {
-			*ddp++ = handle->decoder(handle->buf[y], AUDIO_ENCODING_LINEAR, context);
+			*ddp++ = (int16_t) handle->decoder(handle->buf[y], AUDIO_ENCODING_LINEAR, context);
             new_len += 2;
 		}
 		switch_bitpack_done(&handle->unpack);
